move empty state card component to dir, create external scss, formatting
