+++ app/ui/src/app/common/empty-state-card/empty-state-card.component.scss
@@ -1,0 +1,20 @@
+@import 'syndesis-sass';
+
+:host {
+  display: block;
+
+  .card-pf
+  &:hover {
+    box-shadow: none;
+  }
+
+  .card-pf {
+    background: transparent;
+    padding-bottom: 1px;
+    padding-top: 1px;
+  }
+
+  .card-pf-view {
+    border: 1px dashed $color-pf-black-500;
+  }
+}
